refactor(corebundle): split main.go into 5 topic files (V-A8)
Extract helpers from cmd/corebundle/main.go (423L -> 94L) into dedicated
files while keeping package main and all symbols internal. Pure code move,
zero behavior change.

New files:
- secrets.go: loadSecret / loadKeySet / loadCursorCodec / loadAllCursorCodecs
- adapter_mode.go: validateModeCoupling / validateAdapterMode
- jwt_deps.go: jwtDeps struct + buildJWTDeps
- metrics.go: metricsTokenHeader + withMetricsTokenGuard + promStack helpers
- controlplane.go: internalGuardFromEnv

main.go now holds only package doc, main(), and run().

Refs: PR-A3 V-A8 CMD-THICK-ENTRY-REDUCE-01
Plan: docs/plans/202604232330-025-architecture-pr-implementation-plan.md
